Replaced all types with pairs of terms
The first component represents the type and the second one represents
the sort of the type.

Reason: All valid terms are produced as pairs of the form (term, type).
The types are produced in the same manner. We often need the sort of a
type. Therefore it makes no sense first to produce the types with its
sort, then throw the sort away by pushing the type into the context or
into products and finally recreating the sorts of the types in needed.

Therefore the types in the context, the types in products, lambdas and
let expressions contain the actual type and its sort.